refactor: extract FetchClient class from _fetch_client into _fetch_client_class
Moves the FetchClient class to _fetch_client_class.py, keeping
singleton management and fetch_url wrapper in _fetch_client.py.
